Added a couple of function skeletons for the automatic unloading of unused sounds (coming very soon). Removed the empty functions S_BeginPrecaching, S_EndPrecaching and S_ClearPrecache. I broke the "silentlymissing" sound flag recently, it could be set even if the sound was successfully loaded (this flag doesn't seem to be used in any active part of the code, but until I sort this out it's better to have it right); anyhow it's fixed now. Switched "silentlymissing"'s type to "qboolean" in the process. Fixed an implicit conversion warning when calling Sys_Sleep in MSVC6.
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@3884 d7cf8633-e32d-0410-b094-e92efae38249

+++ host.c
@@ -577,7 +577,7 @@ qboolean Host_FilterTime (double time)
 		{
 			// don't totally hog the CPU
 			if (timeleft >= 0.02)
-				Sys_Sleep(timeleft * 1000 - 5);
+				Sys_Sleep((int)(timeleft * 1000) - 5);
 			return false;
 		}
 	}
